refactor(web): re-home the onboarding overlay into its own module (apps/web/src/onboarding)
