+++ apps/web/src/index.css
@@ -42,7 +42,7 @@
 	--spacing-xl: var(--space-xl);
 
 	/* Onboarding feature-carousel autoplay: the active item's background fills horizontally over 5s
-	   (`forwards` holds it full once complete). Referenced by `panels/Onboarding`. */
+	   (`forwards` holds it full once complete). Referenced by `onboarding/Onboarding`. */
 	--animate-fill: fill 5000ms linear forwards;
 	@keyframes fill {
 		from {
