+++ apps/web/src/onboarding/index.ts
@@ -1,0 +1,1 @@
+export { Onboarding } from "./Onboarding";
