+++ apps/web/src/shell/Shell.tsx
@@ -3,8 +3,8 @@ import { useEffect } from "react";
 import { ErrorBoundary } from "../components/ErrorBoundary";
 import { ResizableHandle, ResizablePanel, ResizablePanelGroup } from "../components/ui/resizable";
 import { PRODUCT_NAME } from "../constants/branding";
+import { Onboarding } from "../onboarding";
 import { CenterTabs } from "../panels/CenterTabs";
-import { Onboarding } from "../panels/Onboarding";
 import { ProjectTree } from "../panels/ProjectTree";
 import { RightPanel } from "../panels/RightPanel";
 import { SettingsDialog } from "../panels/SettingsDialog";
